Dirty test for client.CloseSend() flush
Fairly ugly, quick and dirty test to check that the client.CloseSend()
does not wait until all the messages are flushed, but instead you
need to add a .Sleep(x).

I add here a small loggregator fake server that exposes the received
messages in a channel, and update the main.go to send a bunch of
gauges at the end. Then it waits a customisable amount of seconds
at the end.

There are 2 tests, one with sleep other without.

Theory: I think the problem might be  that client.EmitGauge() is
asynchronous so the client is closed before the metrics are sent.